+++ R/sitrep.R
@@ -33,10 +33,11 @@ check_for_rstudio_updates <- function(os = tolower(Sys.info()[["sysname"]]), ver
     return()
   }
 
-  return(glue::glue('
-    {result[["update-message"]]}.
-    Download at: {ui_field(result[["update-url"]])}
-    ')
+  return(
+    sprintf("%s.\nDownload at: %s",
+      result[["update-message"]],
+      ui_field(result[["update-url"]])
+    )
   )
 }
 
@@ -164,95 +165,4 @@ print.dev_sitrep <- function(x, ...) {
   }
 
   invisible(x)
-}
-
-#' @export
-#' @rdname devtools-deprecated
-dr_devtools <- function() {
-  .Deprecated("dev_sitrep()", package = "devtools")
-  dev_sitrep()
-}
-
-#' @export
-#' @rdname devtools-deprecated
-dr_github <- function(path = ".") {
-  .Deprecated(package = "devtools")
-  if (!uses_git(path)) {
-    return(doctor("github", "Path is not a git repository"))
-  }
-  if (!uses_github(path)) {
-    return(doctor("github", "Path is not a GitHub repository"))
-  }
-
-  msg <- character()
-  r <- git2r::repository(path, discover = TRUE)
-
-  config <- git2r::config(r)
-  config_names <- names(modifyList(config$global, config$local))
-
-  if (!uses_github(path)) {
-    msg[["github"]] <- " * cannot detect that this repo is connected to GitHub"
-  }
-  if (!("user.name" %in% config_names)) {
-    msg[["name"]] <- "* user.name config option not set"
-  }
-  if (!("user.email" %in% config_names)) {
-    msg[["user"]] <- "* user.email config option not set"
-  }
-
-  if (!file.exists("~/.ssh/id_rsa")) {
-    msg[["ssh"]] <- "* SSH private key not found"
-  }
-
-  if (identical(Sys.getenv("GITHUB_PAT"), "")) {
-    msg[["PAT"]] <- paste(
-      "* GITHUB_PAT environment variable not set",
-      "(this is not necessary unless you want to install private repos",
-      "or connect local repos to GitHub)"
-    )
-  }
-
-  desc_path <- file.path(path, "DESCRIPTION")
-  desc <- read_dcf(desc_path)
-  field_empty <- function(d, f) is.null(d[[f]]) || identical(d[[f]], "")
-  field_no_re <- function(d, f, re) !grepl(re, d[[f]])
-
-  re <- "https://github.com/(.*?)/(.*)"
-  if (field_empty(desc, "URL")) {
-    msg[["URL_empty"]] <- "* empty URL field in DESCRIPTION"
-  } else if (field_no_re(desc, "URL", re)) {
-    msg[["URL"]] <- "* no GitHub repo link in URL field in DESCRIPTION"
-  }
-
-  re <- paste0(re, "/issues")
-  if (field_empty(desc, "BugReports")) {
-    msg[["BugReports_empty"]] <- "* empty BugReports field in DESCRIPTION"
-  } else if (field_no_re(desc, "BugReports", re)) {
-    msg[["BugReports"]] <- "* no GitHub Issues link in URL field in DESCRIPTION"
-  }
-
-  doctor("github", msg)
-}
-
-# Doctor class ------------------------------------------------------------
-
-doctor <- function(name, messages) {
-  structure(
-    length(messages) == 0,
-    doctor = paste0("DR_", toupper(name)),
-    messages = messages,
-    class = "doctor"
-  )
-}
-
-#' @export
-print.doctor <- function(x, ...) {
-  if (x) {
-    message(attr(x, "doctor"), " SAYS YOU LOOK HEALTHY")
-    return()
-  }
-
-  warning(attr(x, "doctor"), " FOUND PROBLEMS", call. = FALSE, immediate. = TRUE)
-  messages <- strwrap(attr(x, "messages"), exdent = 2)
-  message(paste(messages, collapse = "\n"))
 }